Add health & damage
arrow & enemy class

+++ Assets/Scripts/ScriptableObjects/ArrowStats.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Arrow Stats", menuName = "ScriptableObjects/GameCore/ArrowStats")]
 public class ArrowStats : ScriptableObject
 {
-    [SerializeField] private float _damage;
+    [SerializeField] private int _damage;
     [SerializeField] private float _speed;
 
-    public float Damage { get => _damage; }
+    public int Damage { get => _damage; }
 
     public float Speed { get => _speed; }
 }
